Dvc git combine (#9)
* s

* initial progress with git-dvc combining process. Testing now.

* changes fixing inference API with git

* some fixes

* add secret to build step env

* testing inference service with API key

* remove if refs tag

* only push one image

* only push test tag

* revert tests

* modified dockerfile, stratup_training, startup_inf, train.sh and the yaml files of inf and train services
An endpoint is added to serve_model that can be called to fetch a certain model given a valid dvc-versioning commit hash (also created shell script which is called by endpoint)
TODO: check if in training service the path to data does not matter when getting data from scrapers. Now the path to the data folder is changed to /app/dvc-versioning/data

* fixed possible bug in startup inf

* fixed some more bugs

* fixed many bugs in startup training.sh, train.sh
some bugs are fixed in inference. TODO: test pull from commit hash and startup

* fixed url bug. TODO: check checkoutandpull sh

* maybe a bug fix

* the new model is now loaded in after pulling it from gdrive using  a commit hash

* updated readme to include github token

* added timeout to give enough time for the checkout commit endpoint to finish. fixed bug in path to shell script needed for checkout commit.

* run isort

* lint

Co-authored-by: M-R-T-U-D <M.Rashad@student.tudelft.nl>